Move createBlastSubmissionData to Submit component
